edmaclog: 70D and 200D support
This splits out some stuff so we can have ARMv7 and ARMv5 code
as required per cam.  The pattern should make extending to other
cams fairly easy.

To support 70D, had to put in a bodge solution for a hang
when using take_semaphore() in a hook context, in some cases.